Refactor imports and improve index page styles.
Updated the PricingTier component import path for consistency and removed unnecessary nested tags in index.vue. Refined text styles for improved readability and alignment with design standards.
