+++ nuxt-ui/components/landing-page/PricingTier.vue
